Refactor formats/art
